Refactor using double instead of float

+++ NAudio.Dsd/Dsd2Pcm/Dither.cs
@@ -3,11 +3,11 @@
     public class Dither
     {
         private readonly Random _random;
-        private readonly float _quantizationStep;
-        private float _previousError1;
-        private float _previousError2;
+        private readonly double _quantizationStep;
+        private double _previousError1;
+        private double _previousError2;
 
-        public delegate float DitherFunction(float input);
+        public delegate double DitherFunction(double input);
         public readonly DitherFunction ApplyDither;
 
         /// <summary>
@@ -46,88 +46,88 @@
         }
 
         #region Dithering Algorithms
-        private float ApplyRectangularDither(float input)
+        private double ApplyRectangularDither(double input)
         {
             // RPDF (Rectangular Probability Density Function)
-            float noise = (_random.NextSingle() - 0.5f) * _quantizationStep;
+            double noise = (_random.NextSingle() - 0.5f) * _quantizationStep;
             return input + noise;
         }
 
-        private float ApplyTriangularPdfDither(float input)
+        private double ApplyTriangularPdfDither(double input)
         {
             // TPDF (Triangular PDF) - sum of two RPDF noises
-            float r1 = (_random.NextSingle() - 0.5f) * _quantizationStep;
-            float r2 = (_random.NextSingle() - 0.5f) * _quantizationStep;
+            double r1 = (_random.NextSingle() - 0.5f) * _quantizationStep;
+            double r2 = (_random.NextSingle() - 0.5f) * _quantizationStep;
             return input + r1 + r2;
         }
 
-        private float ApplyHighPassTriangularPdfDither(float input)
+        private double ApplyHighPassTriangularPdfDither(double input)
         {
             // High-pass filtered TPDF
-            float r1 = (_random.NextSingle() - 0.5f) * _quantizationStep;
-            float r2 = (_random.NextSingle() - 0.5f) * _quantizationStep;
-            float r3 = (_random.NextSingle() - 0.5f) * _quantizationStep;
+            double r1 = (_random.NextSingle() - 0.5f) * _quantizationStep;
+            double r2 = (_random.NextSingle() - 0.5f) * _quantizationStep;
+            double r3 = (_random.NextSingle() - 0.5f) * _quantizationStep;
 
             return input + r1 + r2 - r3;
         }
 
-        private float ApplyModifiedHighPassTriangularPdfDither(float input)
+        private double ApplyModifiedHighPassTriangularPdfDither(double input)
         {
             // Modified high-pass TPDF with different coefficients
-            float r1 = (_random.NextSingle() - 0.5f) * _quantizationStep;
-            float r2 = (_random.NextSingle() - 0.5f) * _quantizationStep;
-            float r3 = (_random.NextSingle() - 0.5f) * _quantizationStep;
+            double r1 = (_random.NextSingle() - 0.5f) * _quantizationStep;
+            double r2 = (_random.NextSingle() - 0.5f) * _quantizationStep;
+            double r3 = (_random.NextSingle() - 0.5f) * _quantizationStep;
 
             return input + 0.75f * r1 + 0.75f * r2 - 0.5f * r3;
         }
 
-        private float ApplyLipshitzMinimallyAudibleDither(float input)
+        private double ApplyLipshitzMinimallyAudibleDither(double input)
         {
             // Lipshitz's minimally audible noise shaping
-            float r1 = (_random.NextSingle() - 0.5f) * _quantizationStep;
-            float r2 = (_random.NextSingle() - 0.5f) * _quantizationStep;
+            double r1 = (_random.NextSingle() - 0.5f) * _quantizationStep;
+            double r2 = (_random.NextSingle() - 0.5f) * _quantizationStep;
 
             // Simple noise shaping
-            float shapedNoise = r1 + r2 - _previousError1;
+            double shapedNoise = r1 + r2 - _previousError1;
             _previousError1 = r1;
 
             return input + shapedNoise;
         }
 
-        private float ApplyFWeightedDither(float input)
+        private double ApplyFWeightedDither(double input)
         {
             // F-weighted noise shaping (approximation)
-            float r1 = (_random.NextSingle() - 0.5f) * _quantizationStep;
-            float r2 = (_random.NextSingle() - 0.5f) * _quantizationStep;
+            double r1 = (_random.NextSingle() - 0.5f) * _quantizationStep;
+            double r2 = (_random.NextSingle() - 0.5f) * _quantizationStep;
 
             // First-order high-pass filter
-            float shapedNoise = r1 + r2 - 0.5f * _previousError1;
+            double shapedNoise = r1 + r2 - 0.5f * _previousError1;
             _previousError1 = shapedNoise;
 
             return input + shapedNoise;
         }
 
-        private float ApplyVanderkooyLipshitzDither(float input)
+        private double ApplyVanderkooyLipshitzDither(double input)
         {
             // Vanderkooy-Lipshitz noise shaping
-            float r1 = (_random.NextSingle() - 0.5f) * _quantizationStep;
-            float r2 = (_random.NextSingle() - 0.5f) * _quantizationStep;
+            double r1 = (_random.NextSingle() - 0.5f) * _quantizationStep;
+            double r2 = (_random.NextSingle() - 0.5f) * _quantizationStep;
 
             // Second-order noise shaping
-            float shapedNoise = r1 + r2 - 1.6f * _previousError1 + 0.64f * _previousError2;
+            double shapedNoise = r1 + r2 - 1.6f * _previousError1 + 0.64f * _previousError2;
             _previousError2 = _previousError1;
             _previousError1 = shapedNoise;
 
             return input + shapedNoise;
         }
 
-        private float ApplySoPdfDither(float input)
+        private double ApplySoPdfDither(double input)
         {
             // Second-order PDF dither
-            float r1 = (_random.NextSingle() - 0.5f) * _quantizationStep;
-            float r2 = (_random.NextSingle() - 0.5f) * _quantizationStep;
-            float r3 = (_random.NextSingle() - 0.5f) * _quantizationStep;
-            float r4 = (_random.NextSingle() - 0.5f) * _quantizationStep;
+            double r1 = (_random.NextSingle() - 0.5f) * _quantizationStep;
+            double r2 = (_random.NextSingle() - 0.5f) * _quantizationStep;
+            double r3 = (_random.NextSingle() - 0.5f) * _quantizationStep;
+            double r4 = (_random.NextSingle() - 0.5f) * _quantizationStep;
 
             return input + r1 + r2 + r3 + r4;
         }
